requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.3.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1

# Document processing
pypdf>=3.17.0
python-docx>=1.0.0
python-pptx>=0.6.21
openpyxl>=3.1.2
pandas>=2.1.0
unstructured>=0.10.0
pdf2image>=1.16.3
pytesseract>=0.3.10

# Web and media handling
youtube-transcript-api>=0.6.1
beautifulsoup4>=4.12.0
requests>=2.31.0

# Vector storage and embeddings
faiss-cpu>=1.7.4
tiktoken>=0.5.1

# Text processing
nltk>=3.8.1

# Optional but recommended for better performance
# uncomment if needed
# chromadb>=0.4.0
# sentence-transformers>=2.2.2

# Development tools
python-magic>=0.4.27
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0 